*: support 32-bit
This change makes the necessary code and test changes to allow
building on 32-bit, and to make all tests pass.

We also add a CI target that uses `GOARCH=386`.

Fixes #2853

+++ batch_test.go
@@ -771,7 +771,7 @@ func TestBatchOpDoesIncrement(t *testing.T) {
 func TestBatchGet(t *testing.T) {
 	testCases := []struct {
 		method       string
-		memTableSize int
+		memTableSize uint64
 	}{
 		{"build", 64 << 20},
 		{"build", 2 << 10},
